# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/roles": {
      "get": {
        "operationId": "RoleController_get",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "all-roles": {
                    "summary": "All Roles",
                    "value": {
                      "data": [
                        {
                          "id": 1,
                          "name": "John Doe",
                          "default": true
                        }
                      ]
                    }
                  },
                  "role-with-id": {
                    "summary": "Role with Id",
                    "value": {
                      "data": [
                        {
                          "id": 1,
                          "name": "John Doe",
                          "default": true,
                          "Permissions": [
                            {
                              "name": "John Doe",
                              "method": [
                                "string"
                              ]
                            }
                          ]
                        }
                      ]
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      },
      "post": {
        "operationId": "RoleController_post",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      }
    },
    "/api/v1/roles/{id}": {
      "get": {
        "operationId": "RoleController_get",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "all-roles": {
                    "summary": "All Roles",
                    "value": {
                      "data": [
                        {
                          "id": 1,
                          "name": "John Doe",
                          "default": true
                        }
                      ]
                    }
                  },
                  "role-with-id": {
                    "summary": "Role with Id",
                    "value": {
                      "data": [
                        {
                          "id": 1,
                          "name": "John Doe",
                          "default": true,
                          "Permissions": [
                            {
                              "name": "John Doe",
                              "method": [
                                "string"
                              ]
                            }
                          ]
                        }
                      ]
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      },
      "patch": {
        "operationId": "RoleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      },
      "delete": {
        "operationId": "RoleController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Aroovia",
    "description": "",
    "version": "1.0",
    "contact": {
      "name": "Fahd Hakem",
      "url": "www.linkedin.com/in/fahd-hakem",
      "email": "fhakem75@gmail.com"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "ACCESS Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "REFRESH Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "PASSWORD_RESET Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "VERIFY Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "VISITOR Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateRoleDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "object",
            "example": "{\"en\": \"John\", \"ar\": \"جون\"}"
          },
          "roleKey": {
            "type": "string"
          },
          "permissionIds": {
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        }
      },
      "CreateRoleDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "object",
            "example": "{\"en\": \"John\", \"ar\": \"جون\"}"
          },
          "roleKey": {
            "type": "string"
          },
          "permissionIds": {
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "name",
          "roleKey"
        ]
      },
...
    }
  }
}